feat: restructured the application
