+++ src/config/index.ts
@@ -1,0 +1,7 @@
+import dotenv from 'dotenv';
+
+dotenv.config();
+
+export const ENV ={
+    MONGO_URL : process.env.MONGO_URL || "",
+}
