+++ src/index.ts
@@ -1,20 +1,17 @@
 import express from "express";
 import dotenv from "dotenv";
-import signup from "./signup";
-import connecttoDB from "./connectToDB";
-import { sign } from "crypto";
+import connecttoDB from "./db/mongo";
+import userRoutes from "./route/userRoutes";
 dotenv.config();
 
 const app = express();
-connecttoDB();
 
 app.use(express.json());
 
-console.log("signup:", signup());
-app.post("/api/v1/signup", signup());
+app.post("/api/v1/user", userRoutes);
 
-app.post("/api/v1/signin", (req, res) => {});
 
+//Follow above approach
 app.get("/api/v1/content", (req, res) => {});
 
 app.delete("/api/v1/content", (req, res) => {});
@@ -24,5 +21,6 @@ app.post("/api/v1/brain/share", (req, res) => {});
 app.post("/api/v1/brain/:sharelink", (req, res) => {});
 
 app.listen(3000, () => {
+  connecttoDB();
   console.log("Server is running on port 3000");
 });
